+++ src/main/ts/core/PreRelease.ts
@@ -1,5 +1,4 @@
 export const releaseCandidate = 'rc';
-export const mainBranch = 'alpha';
 export const featureBranch = 'feature';
 export const hotfixBranch = 'hotfix';
 export const spikeBranch = 'spike';