+++ src/test/ts/commands/StatusTest.ts
@@ -33,32 +33,50 @@ const check = async (dir: string, expected: Object) => {
 };
 
 describe('Status', () => {
-  it('shows status for main', async () => {
+  it('shows status for main branch in preRelease state', async () => {
     const { dir, git } = await newGit();
-    await branchWithPj({ dir, git }, '0.1.0-alpha', 'main');
+    await branchWithPj({ dir, git }, '0.1.0-rc', 'main');
 
     await check(dir, {
-      branchState: 'main',
-      isLatestReleaseBranch: false,
+      branchState: 'releaseCandidate',
+      isLatest: true,
       currentBranch: 'main',
       branchType: 'main',
       version: {
         major: 0,
         minor: 1,
         patch: 0,
-        preRelease: 'alpha'
+        preRelease: 'rc'
       },
-      versionString: '0.1.0-alpha'
+      versionString: '0.1.0-rc'
     });
-  });
+  }).timeout(20000);
 
-  it('shows status for rc', async () => {
+  it('shows status for main branch in releaseReady state', async () => {
+    const { dir, git } = await newGit();
+    await branchWithPj({ dir, git }, '0.7.0', 'main');
+
+    await check(dir, {
+      branchState: 'releaseReady',
+      isLatest: true,
+      currentBranch: 'main',
+      branchType: 'main',
+      version: {
+        major: 0,
+        minor: 7,
+        patch: 0
+      },
+      versionString: '0.7.0'
+    });
+  }).timeout(20000);
+
+  it('shows status for release branch in preRelease state', async () => {
     const { dir, git } = await newGit();
     await branchWithPj({ dir, git }, '1.98.2-rc', 'release/1.98');
 
     await check(dir, {
       branchState: 'releaseCandidate',
-      isLatestReleaseBranch: true,
+      isLatest: true,
       currentBranch: 'release/1.98',
       branchType: 'release',
       version: {
@@ -69,15 +87,15 @@ describe('Status', () => {
       },
       versionString: '1.98.2-rc'
     });
-  });
+  }).timeout(20000);
 
-  it('shows status for releaseReady', async () => {
+  it('shows status for release branch in releaseReady state', async () => {
     const { dir, git } = await newGit();
     await branchWithPj({ dir, git }, '1.98.7', 'release/1.98');
 
     await check(dir, {
       branchState: 'releaseReady',
-      isLatestReleaseBranch: true,
+      isLatest: true,
       currentBranch: 'release/1.98',
       branchType: 'release',
       version: {
@@ -87,24 +105,7 @@ describe('Status', () => {
       },
       versionString: '1.98.7'
     });
-  });
+  }).timeout(20000);
 
-  it('shows status for releaseReady of "old" release', async () => {
-    const { dir, git } = await newGit();
-    await branchWithPj({ dir, git }, '41.98.7', 'release/41.98');
-    await branchWithPj({ dir, git }, '1.100.1', 'release/1.100');
-
-    await check(dir, {
-      branchState: 'releaseReady',
-      isLatestReleaseBranch: false,
-      currentBranch: 'release/1.100',
-      branchType: 'release',
-      version: {
-        major: 1,
-        minor: 100,
-        patch: 1
-      },
-      versionString: '1.100.1'
-    });
-  });
+  // TODO TINY-6924: Add test case where isLatest returns false. Will need to publish, or mock out listing tags.
 });