+++ src/test/ts/core/VersionTest.ts
@@ -3,10 +3,7 @@ import * as chai from 'chai';
 import * as chaiAsPromised from 'chai-as-promised';
 import fc from 'fast-check';
 import * as Version from '../../../main/ts/core/Version';
-import * as PromiseUtils from '../../../main/ts/utils/PromiseUtils';
-
-type Arbitrary<A> = fc.Arbitrary<A>;
-type MajorMinorVersion = Version.MajorMinorVersion;
+import { Comparison } from '../../../main/ts/utils/Comparison';
 
 const assert = chai.use(chaiAsPromised).assert;
 
@@ -117,32 +114,67 @@ describe('Version', () => {
     });
   });
 
-  describe('sortMajorMinorVersions', () => {
-    const arbmm = (): Arbitrary<MajorMinorVersion> => fc.tuple(fc.nat(300), fc.nat(300)).map(([ major, minor ]) => ({ major, minor }));
-
-    it('returns the same length array', () => {
-      fc.assert(fc.property(fc.array(arbmm()), (mms) => {
-        assert.equal(
-          Version.sortMajorMinorVersions(mms).length,
-          mms.length
-        );
+  describe('compareVersions', () => {
+    it('is reflexive for 3-point versions', () => {
+      fc.assert(fc.property(fc.integer(), fc.integer(), fc.integer(), (major, minor, patch) => {
+        assert.equal(Version.compareVersions({ major, minor, patch }, { major, minor, patch }), Comparison.EQ);
       }));
     });
 
-    it('sorts', async () => {
-      const check = async (sInput: string[], sExpected: string[]): Promise<void> => {
-        const mmInput = await PromiseUtils.parMap(sInput, Version.parseMajorMinorVersion);
-        const mmOutput = Version.sortMajorMinorVersions(mmInput);
-        const sOutput = mmOutput.map(Version.majorMinorVersionToString);
-        assert.deepEqual(sOutput, sExpected);
-      };
-      await check([], []);
-      await check([ '3.4' ], [ '3.4' ]);
-      await check([ '1.0', '2.0' ], [ '1.0', '2.0' ]);
-      await check([ '2.0', '1.0' ], [ '1.0', '2.0' ]);
-      await check([ '1.100', '1.10' ], [ '1.10', '1.100' ]);
-      await check([ '1.2', '1.10' ], [ '1.2', '1.10' ]);
-      await check([ '1.10', '1.0' ], [ '1.0', '1.10' ]);
+    it('is reflexive for 3-point versions with prerelease', () => {
+      fc.assert(fc.property(fc.integer(), fc.integer(), fc.integer(), fc.hexaString(), (major, minor, patch, preRelease) => {
+        assert.equal(Version.compareVersions({ major, minor, patch, preRelease }, { major, minor, patch, preRelease }), Comparison.EQ);
+      }));
+    });
+
+    it('considers preRelease versions lower less than release versions of the same 3-point version', () => {
+      fc.assert(fc.property(fc.integer(), fc.integer(), fc.integer(), fc.hexaString(), (major, minor, patch, preRelease) => {
+        assert.equal(Version.compareVersions({ major, minor, patch, preRelease }, { major, minor, patch }), Comparison.LT);
+        assert.equal(Version.compareVersions({ major, minor, patch }, { major, minor, patch, preRelease }), Comparison.GT);
+      }));
+    });
+
+    it('compares major versions', () => {
+      fc.assert(fc.property(fc.integer(0, 1000), fc.integer(), fc.integer(), fc.integer(), fc.integer(), fc.hexaString(), fc.hexaString(),
+        (major, minor1, minor2, patch1, patch2, preRelease1, preRelease2) => {
+          assert.equal(Version.compareVersions(
+            { major, minor: minor1, patch: patch1, preRelease: preRelease1 },
+            { major: major + 1, minor: minor2, patch: patch2, preRelease: preRelease2 }
+          ), Comparison.LT);
+
+          assert.equal(Version.compareVersions(
+            { major: major + 1, minor: minor2, patch: patch2, preRelease: preRelease2 },
+            { major, minor: minor1, patch: patch1, preRelease: preRelease1 }
+          ), Comparison.GT);
+        }));
+    });
+
+    it('compares minor versions', () => {
+      fc.assert(fc.property(fc.integer(), fc.integer(0, 1000), fc.integer(), fc.integer(), fc.hexaString(), fc.hexaString(),
+        (major, minor, patch1, patch2, preRelease1, preRelease2) => {
+          assert.equal(Version.compareVersions(
+            { major, minor, patch: patch1, preRelease: preRelease1 },
+            { major, minor: minor + 1, patch: patch2, preRelease: preRelease2 }
+          ), Comparison.LT);
+          assert.equal(Version.compareVersions(
+            { major, minor: minor + 1, patch: patch2, preRelease: preRelease2 },
+            { major, minor, patch: patch1, preRelease: preRelease1 }
+          ), Comparison.GT);
+        }));
+    });
+
+    it('compares patch versions', () => {
+      fc.assert(fc.property(fc.integer(), fc.integer(), fc.integer(0, 1000), fc.hexaString(), fc.hexaString(),
+        (major, minor, patch, preRelease1, preRelease2) => {
+          assert.equal(Version.compareVersions(
+            { major, minor, patch, preRelease: preRelease1 },
+            { major, minor, patch: patch + 1, preRelease: preRelease2 }
+          ), Comparison.LT);
+          assert.equal(Version.compareVersions(
+            { major, minor, patch: patch + 1, preRelease: preRelease2 },
+            { major, minor, patch, preRelease: preRelease1 }
+          ), Comparison.GT);
+        }));
     });
   });
 });