+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+## [0.12.0] - unreleased
+- publish from main branch
+- tag greatest prerelease version as latest if no prior release versions
+- change output of status command
+
 ## [0.11.0] - 2021-01-11
 ### Added
 - `beehive-flow publish --dist-dir` setting